Add concept sources.
-------------------

Add the concept sources to the repository. They describe several
concepts that will be used in the project. From Lee Graham's Thesis, we
will take the description of the simulation he created for his thesis
and we also write a simulation in OGRE as he did. From Thomas
Geijtenbeek's Thesis, we take the descriptions of the muscles as well as
more of the concepts such as the muscle routing we will use as a changin
parameter to the joint.