Add lines to install.sh script on how to install FFmpeg from the source. Update FindFFMPEG.cmake module with one that adds the include directories excluding the parent folder of the file.
------------

Add lines to install.sh script on how to install FFmpeg from the source
as it is out of the ubuntu repositories. It seems that ubuntu decided in
favor of the libav fork instead of just providing both the original
ffmpeg library and the libav fork with its dev libraries.

Update FindFFMPEG.cmake module with one that adds the include
directories excluding the parent folder of the file. This is important
especially since I use boost and FFmpeg in the same project causing a
header collision between the boost/time.h and the
FFmpeg/libavutil/time.h. The latter does not provide the searched
features and therefore the compilation fails.