Measurements were taken with 0g and 1g.
-------------------------------------------------
The model leg in 0g is in space without touching the ground at any time
during its 100s simulation time. In the plot of its chua circuit
controller, one can see that the limb limits reduce its behavior to a
stable limit cycle. 

With 1g, this is no longer possible, as the model leg hits the ground
and thereby the force is no longer enough to reach the limit cycle. The
movement stops completely after the model leg hits the ground.